change datatype to int64

+++ schedule/charge_detail.go
@@ -2,7 +2,7 @@ package schedule
 
 // ChargeDetail represents charge detail for schedule object.
 type ChargeDetail struct {
-	Amount      int                    `json:"amount"`
+	Amount      int64                  `json:"amount"`
 	Currency    string                 `json:"currency"`
 	Customer    string                 `json:"customer"`
 	Card        *string                `json:"card"`
